// rahul_complete/front/src/app/model/category.ts
export class Parent {
    parent_id?: string;
    parent_name?: string;
}

export class SubMenu {
    parent_id?: string;
    subMenu_id?: string;
    subMenu_name?: string;
}

export class Category {
    parent_id?: string;
    subMenu_id?: string;
    category_id?: string;
    category_name?: string;
}